# sdk/typescript-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactType": {
        "enum": [
          "web_bundle",
          "cli_package",
          "api_service",
          "notebook"
        ],
        "title": "ArtifactType",
        "type": "string"
      },
      "ArtifactUploadResponse": {
        "properties": {
          "artifact_type": {
            "$ref": "#/components/schemas/ArtifactType"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          },
          "submission_id": {
            "format": "uuid",
            "title": "Submission Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "submission_id",
          "artifact_type",
          "storage_key",
          "content_hash",
          "created_at",
          "updated_at"
        ],
        "title": "ArtifactUploadResponse",
        "type": "object"
      },
      "Body_upload_artifact_submissions__submission_id__artifacts_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_artifact_submissions__submission_id__artifacts_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SubmissionValidationResponse": {
        "properties": {
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "submission_id": {
            "format": "uuid",
            "title": "Submission Id",
            "type": "string"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "submission_id",
          "valid",
          "errors"
        ],
        "title": "SubmissionValidationResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Hackathon Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/artifacts/{artifact_id}/download": {
      "get": {
        "operationId": "download_artifact_artifacts__artifact_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/submissions/{submission_id}/artifacts": {
      "post": {
        "operationId": "upload_artifact_submissions__submission_id__artifacts_post",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Submission Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "artifact_type",
            "required": true,
            "schema": {
              "title": "Artifact Type",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_artifact_submissions__submission_id__artifacts_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/submissions/{submission_id}/validation": {
      "get": {
        "operationId": "validate_submission_mandatory_requirements_submissions__submission_id__validation_get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Submission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Submission Mandatory Requirements",
        "tags": [
          "artifacts"
        ]
      }
    }
  }
}